+++ packages/adapter-static/test/apps/spa/svelte.config.cjs
@@ -1,0 +1,9 @@
+/** @type {import('@sveltejs/kit').Config} */
+module.exports = {
+	kit: {
+		adapter: require('../../../index.cjs')({
+			fallback: '200.html'
+		}),
+		target: '#svelte'
+	}
+};