+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sveltejs/kit
 
+## 1.0.0-next.90
+
+### Patch Changes
+
+- ac60208: Exit process after adapting
+
+## 1.0.0-next.89
+
+### Patch Changes
+
+- 927e63c: update the error message of prerender to optionally include the parent variable
+
+## 1.0.0-next.88
+
+### Patch Changes
+
+- 6f2b4a6: Remove references to npm start
+
+## 1.0.0-next.87
+
+### Patch Changes
+
+- 4131467: Prerender fallback page for SPAs
+
+## 1.0.0-next.86
+
+### Patch Changes
+
+- 2130087: Support multiple rel values on anchor tag
+- ba732ff: Report errors in hooks.js
+- a2f3f4b: Rename `start` to `preview` in the CLI and package scripts
+
 ## 1.0.0-next.85
 
 ### Patch Changes